refactor: inline the FAQ activity components into the route file
The rulebook's 'Extract a component only at a real concern boundary'
says a view-only sub-view stays in the route file as local functions —
the maintainer ruled that wording wins over the co-located organ file.
FaqQuestionsAsked, FaqAnswersGiven, and AskerChip are now local,
unexported functions in index.tsx (matching login.tsx's SignInPanel
precedent), their styles merged into ProfileDetail.module.css, and the
-faqActivity files are gone.
